added 3rd test
