+++ applications/CableNetApplication/tests/cable_net_test_case.py
@@ -10,26 +10,12 @@ class CableNetTestCase(KratosUnittest.TestCase):
     '''This class is the basis for the testing the framework
     It can be used to test complete cases with the "CoSimulation-Analysis"
     '''
-
-    def _createTest(self, problem_dir_name, parameter_file_name):
-        self.problem_dir_name = problem_dir_name
-
-        full_parameter_file_name = os.path.join(problem_dir_name, parameter_file_name + '_parameters.json')
-
-        with open(full_parameter_file_name, 'r') as parameter_file:
-            self.cable_net_parameters = KM.Parameters(parameter_file.read())
-
-        # To avoid many prints
-        echo_level = self.cable_net_parameters["problem_data"]["echo_level"].GetInt()
-        if (echo_level == 0):
-            KM.Logger.GetDefaultOutput().SetSeverity(KM.Logger.Severity.WARNING)
-        else:
-            KM.Logger.GetDefaultOutput().SetSeverity(KM.Logger.Severity.INFO)
-
     def _runTest(self):
+        self.test.Run()
         kratos_utils.DeleteTimeFiles(self.problem_dir_name)
     # called only once for this class, opposed of tearDown()
     @classmethod
     def tearDownClass(cls):
         # Cleaning
         kratos_utils.DeleteDirectoryIfExisting("__pycache__")
+
