+++ applications/CableNetApplication/tests/cable_net_test_factory.py
@@ -10,6 +10,7 @@ import KratosMultiphysics.StructuralMechanicsApplication
 import KratosMultiphysics.DEMApplication
 
 from KratosMultiphysics.CoSimulationApplication.co_simulation_analysis import CoSimulationAnalysis
+from KratosMultiphysics.StructuralMechanicsApplication.structural_mechanics_analysis import StructuralMechanicsAnalysis
 
 import cable_net_test_case
 import os
@@ -30,9 +31,23 @@ def GetFilePath(fileName):
 
 class TestCableNetCoSimulationCases(cable_net_test_case.CableNetTestCase):
 
-    def _runTest(self):
-        super(TestCableNetCoSimulationCases,self)._runTest()
-        CoSimulationAnalysis(self.cable_net_parameters).Run()
+    def _createTest(self, problem_dir_name, parameter_file_name):
+        self.problem_dir_name = problem_dir_name
+
+        full_parameter_file_name = os.path.join(problem_dir_name, parameter_file_name + '_parameters.json')
+
+        with open(full_parameter_file_name, 'r') as parameter_file:
+            self.cable_net_parameters = KM.Parameters(parameter_file.read())
+
+        # To avoid many prints
+        echo_level = self.cable_net_parameters["problem_data"]["echo_level"].GetInt()
+        if (echo_level == 0):
+            KM.Logger.GetDefaultOutput().SetSeverity(KM.Logger.Severity.WARNING)
+        else:
+            KM.Logger.GetDefaultOutput().SetSeverity(KM.Logger.Severity.INFO)
+
+        self.test = CoSimulationAnalysis(self.cable_net_parameters)
+
 
     def test_DEMFEMCableNet_SlidingEdges(self):
         if not numpy_available:
@@ -68,6 +83,35 @@ class TestCableNetCoSimulationCases(cable_net_test_case.CableNetTestCase):
             kratos_utils.DeleteDirectoryIfExisting(GetFilePath(dir_name+"/"+project_name+"_MPI_results"))
             kratos_utils.DeleteDirectoryIfExisting(GetFilePath(dir_name+"/"+project_name+"_Post_Files"))
             kratos_utils.DeleteDirectoryIfExisting(GetFilePath(dir_name+"/"+project_name+"_Results_and_Data"))
+
+class TestCableNetFEMCases(cable_net_test_case.CableNetTestCase):
+
+        def _createTest(self, problem_dir_name):
+            self.problem_dir_name = problem_dir_name
+
+            full_parameter_file_name = os.path.join(problem_dir_name,'ProjectParameters.json')
+
+            with open(full_parameter_file_name, 'r') as parameter_file:
+                self.cable_net_parameters = KM.Parameters(parameter_file.read())
+
+            # To avoid many prints
+            echo_level = self.cable_net_parameters["problem_data"]["echo_level"].GetInt()
+            if (echo_level == 0):
+                KM.Logger.GetDefaultOutput().SetSeverity(KM.Logger.Severity.WARNING)
+            else:
+                KM.Logger.GetDefaultOutput().SetSeverity(KM.Logger.Severity.INFO)
+
+            model = KM.Model()
+            self.test = StructuralMechanicsAnalysis(model, self.cable_net_parameters)
+
+
+        def test_SlidingNodes(self):
+            if not numpy_available:
+                self.skipTest("Numpy not available")
+
+            with KratosUnittest.WorkFolderScope(".", __file__):
+                self._createTest("sliding_element")
+                self._runTest()
 
 if __name__ == '__main__':
     KratosUnittest.main()
