+++ applications/CableNetApplication/tests/sliding_element/FEM.mdpa
@@ -1,0 +1,99 @@
+Begin ModelPartData
+//  VARIABLE_NAME value
+End ModelPartData
+
+Begin Properties 0
+End Properties
+Begin Nodes
+    1   0.0   0.0 0.0
+    2   1.0 1.0 0.0
+    3  2.0 0.0 0.0
+    4   1.0 -1.0 0.0
+End Nodes
+
+
+Begin Elements Element3D2N// GUI group identifier: truss
+        1          0          1         2
+        2          0          2         3
+        3          0          3         4
+        4          0          4         1
+End Elements
+
+Begin Conditions PointLoadCondition3D1N
+        1         0    3
+End Conditions
+
+Begin SubModelPart Parts_truss // Group truss // Subtree Parts
+    Begin SubModelPartNodes
+            1
+            2
+            3
+            4
+    End SubModelPartNodes
+    Begin SubModelPartElements
+            1
+            2
+            3
+            4
+    End SubModelPartElements
+    Begin SubModelPartConditions
+    End SubModelPartConditions
+End SubModelPart
+Begin SubModelPart DISPLACEMENT_dirich // Group dirich // Subtree DISPLACEMENT
+    Begin SubModelPartNodes
+            1
+    End SubModelPartNodes
+    Begin SubModelPartElements
+    End SubModelPartElements
+    Begin SubModelPartConditions
+    End SubModelPartConditions
+End SubModelPart
+Begin SubModelPart DISPLACEMENT_dirichXY // Group dirich // Subtree DISPLACEMENT
+    Begin SubModelPartNodes
+            2
+            3
+            4
+    End SubModelPartNodes
+    Begin SubModelPartElements
+    End SubModelPartElements
+    Begin SubModelPartConditions
+    End SubModelPartConditions
+End SubModelPart
+Begin SubModelPart PointLoad3D_Load_on_points_Auto1 // Group dirich // Subtree DISPLACEMENT
+    Begin SubModelPartNodes
+           3
+    End SubModelPartNodes
+    Begin SubModelPartElements
+    End SubModelPartElements
+    Begin SubModelPartConditions
+           1
+    End SubModelPartConditions
+End SubModelPart
+Begin SubModelPart EdgeCablePart1 // Group dirich // Subtree DISPLACEMENT
+    Begin SubModelPartNodes
+            1
+            2
+            3
+            4
+    End SubModelPartNodes
+    Begin SubModelPartElements
+    End SubModelPartElements
+    Begin SubModelPartConditions
+    End SubModelPartConditions
+End SubModelPart
+Begin SubModelPart PrintPart // Group dirich // Subtree DISPLACEMENT
+    Begin SubModelPartNodes
+            1
+            2
+            3
+            4
+    End SubModelPartNodes
+    Begin SubModelPartElements
+        1
+        2
+        3
+    End SubModelPartElements
+    Begin SubModelPartConditions
+    1
+    End SubModelPartConditions
+End SubModelPart
