+++ applications/CableNetApplication/tests/sliding_element/StructuralMaterials.json
@@ -1,0 +1,38 @@
+{
+    "properties" : [{
+        "model_part_name" : "Structure.Parts_truss",
+        "properties_id"   : 1,
+        "Material"        : {
+            "constitutive_law" : {
+                "name" : "TrussConstitutiveLaw"
+            },
+            "Variables"        : {
+                "DENSITY"             : 7850.0,
+                "YOUNG_MODULUS"       : 10000,
+                "CROSS_AREA"          : 0.1,
+                "TRUSS_PRESTRESS_PK2" : 0.0,
+                "RAYLEIGH_ALPHA"      :  1.1,
+                "RAYLEIGH_BETA"       :  0.000001
+            },
+            "Tables"           : {}
+        }
+    },
+    {
+        "model_part_name" : "Structure.EdgeCablePart1",
+        "properties_id"   : 2,
+        "Material"        : {
+            "constitutive_law" : {
+                "name" : "TrussConstitutiveLaw"
+            },
+            "Variables"        : {
+                "DENSITY"             : 7850.0,
+                "YOUNG_MODULUS"       : 100000000,
+                "CROSS_AREA"          : 0.1,
+                "TRUSS_PRESTRESS_PK2" : 0.0,
+                "RAYLEIGH_ALPHA"      :  20,
+                "RAYLEIGH_BETA"       :  0.00001
+            },
+            "Tables"           : {}
+        }
+    }]
+}
