+++ applications/CableNetApplication/tests/sliding_element/sliding_node_results.json
@@ -1,0 +1,1 @@
+{"TIME": [0.10000000000000007, 0.20000000000000015, 0.3000000000000002, 0.4000000000000003, 0.5000000000000003, 0.6000000000000004, 0.7000000000000005, 0.8000000000000006, 0.9000000000000007, 1.0010000000000006, 1.1019999999999894, 1.2029999999999783, 1.3039999999999672, 1.404999999999956, 1.505999999999945, 1.6069999999999338, 1.7079999999999227, 1.8089999999999116, 1.9099999999999004, 2.0109999999998895, 2.1119999999998784, 2.2129999999998673, 2.313999999999856, 2.414999999999845, 2.515999999999834, 2.616999999999823, 2.7179999999998117, 2.8189999999998006, 2.9199999999997894, 3.0209999999997783, 3.121999999999767, 3.222999999999756, 3.323999999999745, 3.424999999999734, 3.5259999999997227, 3.6269999999997116, 3.7279999999997004, 3.8289999999996893, 3.929999999999678, 4.0299999999996805, 4.129999999999714, 4.229999999999747, 4.329999999999781, 4.429999999999814, 4.5299999999998475, 4.629999999999881, 4.729999999999914, 4.829999999999948, 4.929999999999981, 5.0300000000000145, 5.130000000000048, 5.230000000000081, 5.330000000000115, 5.430000000000148, 5.530000000000181, 5.630000000000215, 5.730000000000248, 5.830000000000282, 5.930000000000315], "NODE_1": {"DISPLACEMENT_X": [0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0, 0.0]}, "NODE_2": {"DISPLACEMENT_X": [1.5742494991876648e-08, 5.810214515790338e-07, 4.304811091556584e-06, 1.6845262797958358e-05, 4.6983650080647036e-05, 0.00010638301301407842, 0.0002092059321087497, 0.0003716154161761478, 0.0006111644348439758, 0.000949978228786721, 0.0014047473279075527, 0.001993139393775173, 0.0027309250333991666, 0.0036311104346275487, 0.00470314991234618, 0.005952287077712706, 0.007379069255656838, 0.008979071935082445, 0.01074285881014319, 0.01265618896755966, 0.014700466826190656, 0.016853413606732932, 0.019089922711637577, 0.021383046951560486, 0.023705054685954262, 0.026028486149757588, 0.02832714158814086, 0.030576939653055617, 0.03275659721443145, 0.0348480987217464, 0.036836942187383856, 0.03871216709371977, 0.04046618456389356, 0.04209444007923784, 0.04359494274644893, 0.04496769214339799, 0.04621402383261273, 0.04733587681386187, 0.048334957523238756, 0.049203650521828964, 0.04995040685269105, 0.05056884640495253, 0.05104572373914913, 0.052067880088891856, 0.056566779255388035, 0.05826687349163579, 0.059808481039127434, 0.06088538958975999, 0.061667401064534555, 0.062240268147863334, 0.06264878901452677, 0.06294361085264115, 0.06315511684444959, 0.06330774056105022, 0.0634180098707053, 0.06349818982971693, 0.06355680558298936, 0.063599961026005, 0.06363198296279]}, "NODE_3": {"DISPLACEMENT_X": [0.0006984079733719172, 0.0038603209544941955, 0.009936808598114116, 0.018955861150722085, 0.030875805404148835, 0.04563223712375002, 0.06314518267416132, 0.08332146233192458, 0.10605672472656905, 0.13150139233066357, 0.15931722037007975, 0.18937624340245693, 0.22154879543511122, 0.25570613812222587, 0.2917230345151837, 0.32948022929050463, 0.368866808849134, 0.40978242641166895, 0.4521393869249434, 0.4958645935400638, 0.5409013617134321, 0.5872111095777324, 0.6347749358442042, 0.6835951013405301, 0.7336964396936808, 0.7851277386975203, 0.8379631580651465, 0.8923037824631886, 0.9482794516155073, 1.006051063063267, 1.0658136109621987, 1.1278003128844298, 1.1922882990557417, 1.2596065180586478, 1.3301467897929284, 1.404379379419157, 1.4828751969851452, 1.5663379693216046, 1.655651906665212, 1.750962064497211, 1.8545803348207086, 1.9684932981888479, 2.0954803659604933, 2.235247938340651, 2.213880294838601, 2.2427082908027796, 2.2437249350171062, 2.2518935240522473, 2.256993581178481, 2.258919649241228, 2.261204084809892, 2.2626077012591623, 2.2636764230534476, 2.2643035340677065, 2.265213612923277, 2.2655467924610577, 2.2662789130802308, 2.2665442217841414, 2.267215978054487]}, "NODE_4": {"DISPLACEMENT_X": [0.0003131117499113643, 0.0018641233254427523, 0.004896137491944707, 0.009463785952570159, 0.015607930927421689, 0.02337777071226713, 0.032833290085437256, 0.04404498758069862, 0.057093156217529056, 0.07222689881481378, 0.08942575784159036, 0.10879831647333561, 0.130460613651404, 0.1545357950002084, 0.18115395159011943, 0.210452152686557, 0.24257466915282166, 0.277673378508734, 0.3159083418387719, 0.3574485473403707, 0.40247282533110795, 0.45117095455393363, 0.5037449988784072, 0.5604109361619463, 0.6214006665308964, 0.6869645157684273, 0.757374382141556, 0.8329277149061368, 0.9139525653383096, 1.0008140251020958, 1.0939224751532437, 1.1937442308571764, 1.3008154158099372, 1.415760276360279, 1.5393157417038208, 1.6723649807109644, 1.8159842590945283, 1.971510033987442, 2.1406378723466495, 2.323656038657809, 2.525032910092827, 2.7487361913631196, 3.000563228156754, 3.233593483137696, 3.2063526216136773, 3.2385522615951223, 3.2462323511801, 3.251772542013596, 3.2559085588637697, 3.2590693379061344, 3.261535576985946, 3.2625019868170826, 3.263777692933243, 3.264668467275911, 3.2649714129159357, 3.265927544023669, 3.2660406319737456, 3.26694081707587, 3.2669768744555565]}, "ELEMENT_5": {}}
